Every known object is being add, updated and deleted
The insert json has all the information to insert a new object into the working memory.

The update json has to contain the exact information about the Object at time.
Example:
If a Person was inserted with name Paul and temperature 37 at first, when its been updated, the user has to provide a json with Person name equals to Paul and update its temperature to 38.
Else the update will change all the not informed values and assign them a null value.

To delete the user only has to provide the object id with its type.

Closes #7